Separate result page Majors component
